edx-installation/scripts/requirements.txt
pycryptodomex==3.9.8
pyjwkest==1.4.2
